refactor: extract new files into a specific folder
