+++ packages/core/src/App/Components/Layout/Header/dtrader-v2/header-account-actions-dtrader-v2.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import { observer, useStore } from '@deriv/stores';
-import { AccountActionsDTraderV2 } from 'App/Components/Layout/Header';
-import { getCurrencyName } from '@deriv/shared';
-import { Localize } from '@deriv/translations';
+import { AccountActionsDTraderV2 } from './account-actions-dtrader-v2';
+import { getAccountTitle } from './Utils/account-switcher-dtrader-v2-utils';
 
 const HeaderAccountActionsDTraderV2 = observer(() => {
     const { client, ui, notifications } = useStore();
@@ -14,37 +13,11 @@ const HeaderAccountActionsDTraderV2 = observer(() => {
         toggleAccountsDialog,
     } = ui;
     const { notifications: notifications_array } = notifications;
-    //TODO: move to helper (together with the duplicates)
-    const currencyDisplay = ({
-        currency,
-        loginid,
-        is_virtual,
-    }: {
-        currency?: string;
-        loginid?: string;
-        is_virtual?: boolean;
-    }) => {
-        const account_type = loginid?.replace(/\d/g, '');
-
-        if (account_type === 'MF') {
-            return <Localize i18n_default_text='Multipliers' />;
-        }
-
-        if (is_virtual) {
-            return <Localize i18n_default_text='Demo' />;
-        }
-
-        if (!currency) {
-            return <Localize i18n_default_text='No currency assigned' />;
-        }
-
-        return getCurrencyName(currency);
-    };
 
     return (
         <div className='header-v2__acc-info__container'>
             <AccountActionsDTraderV2
-                account_switcher_title={currencyDisplay({ currency, loginid, is_virtual })}
+                account_switcher_title={getAccountTitle({ currency, loginid, is_virtual })}
                 acc_switcher_disabled_message={account_switcher_disabled_message}
                 account_type={account_type}
                 balance={balance}
