+++ packages/core/src/App/Components/Layout/Header/index.js
@@ -3,6 +3,5 @@ import PlatformSwitcher from './platform-switcher.jsx';
 
 export * from './menu-links.jsx';
 export * from './account-actions.jsx';
-export * from './account-actions-dtrader-v2';
 export * from './wallets/account-actions-wallets';
 export { ToggleNotifications, PlatformSwitcher };
